ContextAgent_Script/requirements.txt
numpy
anthropic
langgraph
typing_extensions
langchain_community
#langchain_openai
#langchain_google_genai
langchain_experimental
langchain
msticpy[vt3]
langchainhub
#langchain-google-genai
nest-asyncio
grpcio
#langchain-openai
python-dotenv
#msticpy
requests
#openai
langchain_ollama
fastapi 
uvicorn
nats-py